change location of lduMatrix based solvers
