+++ CG/GKOCG.C
